refactor #26: 컬러,이미지 토글버튼 포함한 컴포넌트로 분리
- ToggleOption컴포넌트로 배경을 컬러 또는 이미지 중에서 옵션을 선택하는 기능을 분리함
