+++ src/styles/create-message-form/ToggleOption.scss
@@ -1,0 +1,33 @@
+@import '../variables', '../reset';
+
+@mixin flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.background-form__radio-group {
+  display: flex;
+  margin: 24px 0 28px 0;
+}
+.background-form__radio {
+  display: none;
+}
+.background-form__radio-label {
+  cursor: pointer;
+  @include flex-center;
+  width: 118px;
+  height: 40px;
+  padding: 7px 14px;
+  border-radius: 6px;
+  @include font-style(16, Regular);
+  text-align: center;
+  border: none;
+  background-color: $gray100;
+}
+.background-form__radio--active {
+  @include font-style(16, Bold);
+  border: 2px solid $accent-background-thick;
+  color: $accent-color;
+  background-color: $accent-background-light;
+}
